requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
oauthlib==3.1.0
PySocks==1.7.1
six==1.15.0
websockets==8.1
yarl==1.5.1